requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
pymilvus==2.3.1
PyMuPDF==1.23.8
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7 